Relocate EtwWriteSyncer.go to zapetw module for improved organization
- Renamed and moved cni/log/ETWZapCore/EtwWriteSyncer.go to zapetw/write_syncer.go.
